requirements.txt
Django>=3.2,<3.3
wagtail>=3,<3.1
django_libsass
django-compressor<3
psycopg2-binary
wagtail-markdown<1
django-countries>=7,<8
wagtail-localize>=1.2,<1.3
wagtail-footnotes>=0.8.1,<0.10
lxml>=4.7
sorl-thumbnail
uTidylib==0.8
